requirements.txt
fastapi
uvicorn
pyjwt
aiosmtplib
asyncpg
jose
sqlalchemy
psycopg2
passlib